+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Bound the release-date control by version so it cannot certify a heading it does not recognise ([pm-linear-ndrd](https://github.com/unbraind/pm-linear/blob/main/.agents/pm/issues/pm-linear-ndrd.toon))
 - Adopt pm CLI 2026.9.5 and repair the release-date control that its changelog fix invalidated ([pm-linear-tfj7](https://github.com/unbraind/pm-linear/blob/main/.agents/pm/chores/pm-linear-tfj7.toon))
 
 ## 2026.9.1 - 2026-09-01